iTerm: use variable for dynamic profile config
This is more complicated (yuck) but doing it to make the dependency
among the different places here totally explicit.

It may be a smell to define the variable inside the "iterm" role but
then reference it from outside in the "dotfiles" role, but I am not sure
that the alternative of hoisting it up to the global level is much
better or clearer. At least with this approach we have *some* locality.