vim: Fix edge cases with J/K in visual mode
- Restrict to only VISUAL LINE mode.
- Don't drop selection and break the world at the top and bottom of the
  buffer.

Note the `s:AtBottom` and `s:AtTop` functions that I am not actually
using right now. They work in Neovim. In Vim, the `s:AtBottom` one is
busted because `line('.')` is always the first line in the selection,
not the last one, despite what the docs indicate (note: I am looking at
the Neovim docs, so maybe that is why).

In any case, I found a simpler version that works in both based on
`a:firstline` and `a:lastline`, so I am going to use that. Just
including the unused functions in this commit so it gets reflected in
the historical record.