karabiner: Add back SpaceFN, again
This one is sooo close to actually working like the old one used to, and
is possible thanks to the recently added `to_if_held_down`.

- Press and hold Space for at least 500ms and the SpaceFN layer gets
  activated, at which point you can:
  - Use (Colemak) H/J/K/L for cursor movement.
  - Use (Qwerty) I/J/K/L (ie. inverted T-shape) for cursor movement;
    actual (Colemak) keys for that are U/N/E/I, which conveniently don't
    overlap with H/J/K/L.
  - Use B for space (with optional repeat); this one is in there because
    we lose key repeat on Space because of the use of `to_if_held_down`
    (and in any case, I am using `repeat: false` in order to reduce the
    number of unwanted space events that we get in the 500ms while
    waiting for the layer to activate from several to just one; see
    below).

Only imperfection now is that if you press-and-hold to activate the
mode, you get an immediate space key-down/key-up. Ideally that would be
delayed until key up. So to have this work really well it looks like
we still need a `__BlockUntilKeyUp__` equivalent; see:

https://github.com/tekezo/Karabiner-Elements/issues/877

Nevertheless, despite the fact that it's imperfect, it still seems good
enough to keep, and it is pretty hard to accidentally activate this.
Still probably want to tweak this to get modifiers working nicely as
well; at the moment you can't fake things like Shift-Left with the layer
yet.